+++ Runtime/Groups/AudioGroupSettings.cs
@@ -27,14 +27,14 @@ namespace ActionCode.Audio
         public AudioMixer Mixer => mixer;
 
         /// <summary>
-        /// The Volume in linear range (0F -> 1F).
+        /// The Volume in linear range (0 -> 100).
         /// </summary>
-        public float Volume
+        public uint Volume
         {
             get
             {
                 var hasVolumeParam = mixer.GetFloat(volumeParamName, out float volume);
-                return hasVolumeParam ? DecibelToLinear(volume) : 0F;
+                return hasVolumeParam ? DecibelToLinear(volume) : 0;
             }
             set
             {
@@ -43,8 +43,8 @@ namespace ActionCode.Audio
             }
         }
 
-        private const float noVolumeLN = 0F;
-        private const float defaultVolumeLN = 1F;
+        private const uint noVolume = 0;
+        private const uint defaultVolume = 100;
         private const float maxVolumeDB = 20F;
         private const float minVolumeDB = -80F;
 
@@ -52,21 +52,28 @@ namespace ActionCode.Audio
         /// Checks if the Audio Group has Volume.
         /// </summary>
         /// <returns>Whether the Audio Group has Volume.</returns>
-        public bool HasVolume() => !Mathf.Approximately(Volume, noVolumeLN);
+        public bool HasVolume() => Volume == noVolume;
 
         /// <summary>
         /// Resets the <see cref="Volume"/> to its default value.
         /// </summary>
-        public void ResetVolume() => Volume = defaultVolumeLN;
+        public void ResetVolume() => Volume = defaultVolume;
 
-        private static float DecibelToLinear(float value) => Mathf.Pow(10F, value / maxVolumeDB);
-
-        private static float LinearToDecibel(float value)
+        private static uint DecibelToLinear(float value)
         {
-            var hasValue = value > 0F;
-            return hasValue ?
-                Mathf.Log10(value) * maxVolumeDB :
+            var volumeDB = Mathf.Pow(10F, value / maxVolumeDB);
+            return (uint)(volumeDB * 100);
+        }
+
+        private static float LinearToDecibel(uint value)
+        {
+            var hasValue = value > noVolume;
+            var valuDB = value * 0.01F;
+            var a = hasValue ?
+                Mathf.Log10(valuDB) * maxVolumeDB :
                 minVolumeDB;
+
+            return a;
         }
     }
 }